Move stable sorter to implementation file
Only used by non templated ArrayBase class
